chore: debug requires own namespace
